+++ app-ghc-8.2/GHCMain.hs
@@ -74,6 +74,8 @@ import Data.Char
 import Data.List
 import Data.Maybe
 
+import qualified GHC.Paths
+
 -----------------------------------------------------------------------------
 -- ToDo:
 
@@ -109,7 +111,7 @@ ghcMain setting = do
     argv0 <- getArgs
 
     let (minusB_args, argv1) = partition ("-B" `isPrefixOf`) argv0
-        mbMinusB | null minusB_args = Nothing
+        mbMinusB | null minusB_args = Just GHC.Paths.libdir
                  | otherwise = Just (drop 2 (last minusB_args))
 
     let argv2 = map (mkGeneralLocated "on the commandline") argv1