+++ Changelog.md
@@ -1,4 +1,9 @@
 
+20180121 haskell-dap-0.0.2.0
+
+  * [MODIFY] using GHC.Paths module to set libdir.
+
+
 20180101 haskell-dap-0.0.1.0
 
   * [INFO] Initial release.